% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{add_pid_to_tax}
\alias{add_pid_to_tax}
\title{Add PID and bootstrap values to tax result.}
\usage{
add_pid_to_tax(tax_result, pid)
}
\arguments{
\item{pid}{}
}
\description{
Add PID and bootstrap values to tax result.
}
