% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_squares.R
\name{add_squares}
\alias{add_squares}
\title{Add two squares}
\usage{
add_squares(x, y)
}
\arguments{
\item{x}{A number of vector or logical}

\item{y}{Anumber or vector or logical}
}
\value{
The sum of \code{x}^2 and \code{x}^2
}
\description{
Square two numbers or vectors and take the sum
}
\details{
This function take two numbers...Here are many details
}
\examples{
add_squares(1, -2)

}
\author{
Concillia Mpofu
}
