% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{assign_taxonomyDada2}
\alias{assign_taxonomyDada2}
\title{Assign taxonomy}
\usage{
assign_taxonomyDada2(
  abund_asv_table,
  ref_database,
  taxresults_file,
  minBoot = 0,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE
)
}
\arguments{
\item{taxresults_file}{}
}
\description{
Assign taxonomy
}
