% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{assignTax_as_char}
\alias{assignTax_as_char}
\title{Combine taxonomic assigns and bootstrap values for each locus into single classfication vector}
\usage{
assignTax_as_char(tax_results)
}
\arguments{
\item{tax_results}{}
}
\description{
Combine taxonomic assigns and bootstrap values for each locus into single classfication vector
}
