% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{assignTax_function}
\alias{assignTax_function}
\title{Assign rps10 and ITS taxonomy}
\usage{
assignTax_function(
  raw_path,
  database_rps10,
  database_its,
  data_tables,
  asv_abund_matrix,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE,
  is_ITS = FALSE
)
}
\arguments{
\item{data_tables}{}

\item{asv_abund_matrix}{}

\item{tryRC}{}

\item{verbose}{}

\item{multithread}{(Optional). Default is FALSE.
If TRUE, multithreading is enabled and the number of available threads is automatically determined.   
If an integer is provided, the number of threads to use is set by passing the argument on to
\code{\link{setThreadOptions}}.}

\item{is_ITS}{TRUE or FALSE}
}
\description{
Assign rps10 and ITS taxonomy
}
