% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vcf_GT.R
\name{compare_vcfs_GT}
\alias{compare_vcfs_GT}
\title{Compare genotypes from two vcf files}
\usage{
compare_vcfs_GT(vcf1, vcf2)
}
\arguments{
\item{vcf1}{first vcf file}

\item{vcf2}{second vcf file}
}
\value{
A matrix of logical
}
\description{
Compare genotypes from two vcf files
}
