% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_databases.R
\name{create_ref_database}
\alias{create_ref_database}
\title{Prepare reference databases folder if it doesn't exist.}
\usage{
create_ref_database(intermediate_path)
}
\arguments{
\item{intermediate_path}{}
}
\value{
A new directory called reference_databases
}
\description{
Prepare reference databases folder if it doesn't exist.
}
