% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{cut_trim}
\alias{cut_trim}
\title{Main command to trim primers based on DADA2 functions}
\usage{
cut_trim(returnList, intermediate_path, cutadapt_path)
}
\arguments{
\item{intermediate_path}{}

\item{cutadapt_path}{}
}
\description{
Main command to trim primers based on DADA2 functions
}
