% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cutadapt_functions.R
\name{cutadapt_run}
\alias{cutadapt_run}
\title{Core function for running cutadapt}
\usage{
cutadapt_run(cutadapt_path, cutadapt_data, min_length = 50)
}
\arguments{
\item{min_length}{Default is 50.}
}
\description{
Core function for running cutadapt
}
