% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim_functions.R
\name{cutadapt_tibble}
\alias{cutadapt_tibble}
\title{Prepare for primmer trimming with Cutaapt. Make new sub-directories and specify paths for the trimmed and untrimmed reads}
\usage{
cutadapt_tibble(fastq_data, metadata, intermediate_path)
}
\arguments{
\item{fastq_data}{A path to FASTQ files for analysis}

\item{metadata}{A metadata containing the concatenated metadata and primer data}

\item{intermediate_path}{A path to the intermediate folder and directory}
}
\description{
Prepare for primmer trimming with Cutaapt. Make new sub-directories and specify paths for the trimmed and untrimmed reads
}
