% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{dada2_readcounts_multi_sample}
\alias{dada2_readcounts_multi_sample}
\title{Final inventory of read counts after each step from input to removal of chimeras. This function deals with if you have more than one sample. TODO optimize for one sample}
\usage{
dada2_readcounts_multi_sample(asv_abund_matrix)
}
\arguments{
\item{asv_abund_matrix}{An abundance matrix containing amplified sequence variants}
}
\description{
Final inventory of read counts after each step from input to removal of chimeras. This function deals with if you have more than one sample. TODO optimize for one sample
}
