% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalkulacky.R
\name{dochodkova_kalkulacka}
\alias{dochodkova_kalkulacka}
\title{Dochodkova kalkulacka}
\usage{
dochodkova_kalkulacka(m, n, r1, r2, X)
}
\arguments{
\item{m}{O kolko rokov chce ist do dochodku}

\item{n}{Kolko rokov chce poberat dochodok}

\item{r1}{Urokova miera po zaciatok vyberania dochodku v percentach}

\item{r2}{Urokova miera po zacati vyberiania dochodku v percentach}

\item{X}{Kolko eur chce poberat rocne dochodok}
}
\value{
Kolko si treba rocne odlozit
}
\description{
Vypocita, kolko si musi clovek rocne odlozit, podla toho, kolko chce poberat rocny dochodok
}
\examples{
dochodok <- dochodkova_kalkulacka(40,20,5,5,12000);
}
