% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind_doc.R
\name{fbind}
\alias{fbind}
\title{Concatenate Factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{Factor, A vector of factor levels}

\item{b}{Factor, A vector of factor levels}
}
\value{
A vector of factor levels
}
\description{
Concatenate Factors
}
\examples{
fbind(as.factor(c("cat", "elephant")), as.factor(c("dog", "fish")))
}
