% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_GT_per_sample.R
\name{find_unique_GT_per_sample}
\alias{find_unique_GT_per_sample}
\title{Find unique genotype for each sample from vcf file}
\usage{
find_unique_GT_per_sample(vcf_file)
}
\arguments{
\item{vcf_file}{a vcf file}
}
\value{
a csv file for each sample with unique genotype
}
\description{
Find unique genotype for each sample from vcf file
}
