% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_databases.R
\name{format_database2}
\alias{format_database2}
\title{An ITS database that has modified headers and is output in the reference_databases folder.}
\usage{
format_database2(raw_data_path, database_name)
}
\arguments{
\item{database_name}{}
}
\value{
A rps10 database that has modified headers and is output in the reference_databases folder.
}
\description{
An ITS database that has modified headers and is output in the reference_databases folder.
}
