% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{get_fastq_paths}
\alias{get_fastq_paths}
\title{Retrieve the paths of the filtered and trimmed Fastq ifiles}
\usage{
get_fastq_paths(my_direction, my_primer_pair_id)
}
\arguments{
\item{cutadapt_data}{}
}
\description{
Retrieve the paths of the filtered and trimmed Fastq ifiles
}
