% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Files_Primers.R
\name{get_post_primer_hits}
\alias{get_post_primer_hits}
\title{Get primer counts fo reach sample after primer removal and trimming steps}
\usage{
get_post_primer_hits(primer_data, cutadapt_data, intermediate_path)
}
\arguments{
\item{primer_data}{The primer data tibble created in orient_primers function}

\item{cutadapt_data}{Intermediate_data folder with trimmed and filtered reads for each sample}

\item{intermediate_path}{A path to the intermediate folder and directory}
}
\description{
Get primer counts fo reach sample after primer removal and trimming steps
}
