% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalkulacky.R
\name{hypotekarna_kalkulacka}
\alias{hypotekarna_kalkulacka}
\title{Hypotekarna kalkulacka}
\usage{
hypotekarna_kalkulacka(urok, dlzka, hodnota)
}
\arguments{
\item{r}{Vyska urokovej miery v percentach}

\item{n}{Dlzka splacania v rokoch}

\item{H}{Pozicana suma}
}
\value{
Vyska mesacnej splatky
}
\description{
Vyrata vysku mesacnej splatky podla urokovej miery, dlzky splacania a pozicanej sumy
}
\examples{
splatka1 <- hypotekarna_kalkulacka(5, 30, 100000);
splatka2 <- hypotekarna_kalkulacka(4.5, 20 ,75000);
}
