% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{ITS_rps10_barcode_function}
\alias{ITS_rps10_barcode_function}
\title{Main trim command to run core DADA2 functions for 2 barcodes}
\usage{
ITS_rps10_barcode_function(intermediate_path, cutadapt_data)
}
\arguments{
\item{intermediate_path}{A path to the intermediate folder and directory}

\item{cutadapt_data}{Intermediate_data folder with trimmed and filtered reads for each sample}
}
\description{
Main trim command to run core DADA2 functions for 2 barcodes
}
