% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_anno.R
\name{load_anno}
\alias{load_anno}
\title{Load annotation data from gene ontology database}
\usage{
load_anno(orgdb, keytype, ont, drop)
}
\arguments{
\item{orgdb}{package with sepcific species from AnnotationDbi}

\item{keytype}{ID class}

\item{ont}{ontology "BP" or "MF" or "CC"}

\item{drop}{logical value, TRUE means drop IEA evidence code}
}
\value{
a list contains four lists
}
\description{
Load annotation data from gene ontology database
}
\examples{
tmp <- load_anno(org.Hs.eg.db, "ENTREZID", "BP", TURE)

}
