% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{main_cutadapt_function}
\alias{main_cutadapt_function}
\title{Runs all of the functions associated with running cutadapt. Users should not have to touch these functions.}
\usage{
main_cutadapt_function(
  directory_path,
  primer_path,
  metadata_path,
  fastq_path,
  intermediate_path,
  cutadapt_path
)
}
\arguments{
\item{cutadapt_path}{}
}
\value{
Intermediate_data folder with trimmed and filtered reads for each sample.
}
\description{
Runs all of the functions associated with running cutadapt. Users should not have to touch these functions.
}
