% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{make_abund_table}
\alias{make_abund_table}
\title{Quality filtering to remove chimeras and short sequences}
\usage{
make_abund_table(
  raw_seqtab,
  method = "consensus",
  verbose = FALSE,
  min_asv_length = 50
)
}
\arguments{
\item{method}{(Optional). Default is "consensus". Only has an effect if a sequence table is provided.

  If "pooled": The samples in the sequence table are all pooled together for bimera
     identification (\code{\link[dada2]{isBimeraDenovo}}).   
     
  If "consensus": The samples in a sequence table are independently checked for bimeras,
     and a consensus decision on each sequence variant is made (\code{\link[dada2]{isBimeraDenovoTable}}).

  If "per-sample": The samples in a sequence table are independently checked for bimeras,
     and sequence variants are removed (zeroed-out) from samples independently (\code{\link[dada2]{isBimeraDenovo}}).}

\item{verbose}{(Optional). Default FALSE. 
Print verbose text output.}

\item{min_asv_length}{}
}
\description{
Quality filtering to remove chimeras and short sequences
}
