% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{make_asvAbund_matrix}
\alias{make_asvAbund_matrix}
\title{Make an amplified sequence variants abundance matrix with data processed through preceding steps}
\usage{
make_asvAbund_matrix(
  returnList,
  intermediate_path,
  rawSeqTab_fileName = "Seqcounts_raw_abund_matrix.pdf",
  abundMatrix_fileName = "Seqcounts_postfiltered_abund_matrix.pdf"
)

make_asvAbund_matrix(
  returnList,
  intermediate_path,
  rawSeqTab_fileName = "Seqcounts_raw_abund_matrix.pdf",
  abundMatrix_fileName = "Seqcounts_postfiltered_abund_matrix.pdf"
)
}
\arguments{
\item{returnList}{A list containing data modified by cutadapt, primer data, FASTQ data, and concatenated metadata and primer data}

\item{rawSeqTab_fileName}{A filename as which the raw sequence table will be saved}

\item{abundMatrix_fileName}{A filenmae as which the abundance matrix will be saved}
}
\value{
asv_df

asv_df
}
\description{
Make an amplified sequence variants abundance matrix with data processed through preceding steps

Make an amplified sequence variants abundance matrix with data processed through preceding steps
}
\examples{
asv_abund_matrix <- make_asvAbund_matrix(returnList)
asv_abund_matrix <- make_asvAbund_matrix(returnList)
}
