% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{makeSeqtab}
\alias{makeSeqtab}
\title{Make ASV sequence table}
\usage{
makeSeqtab(merged_reads, orderBy = "abundance")
}
\arguments{
\item{merged_reads}{}

\item{orderBy}{(Optional). \code{character(1)}. Default "abundance".
Specifies how the sequences (columns) of the returned table should be ordered (decreasing).
Valid values: "abundance", "nsamples", NULL.}
}
\description{
Make ASV sequence table
}
