% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{merge_reads_command}
\alias{merge_reads_command}
\title{Merge forward and reverse reads ###CHECK}
\usage{
merge_reads_command(
  intermediate_path,
  minOverlap = 12,
  maxMismatch = 0,
  returnRejects = FALSE,
  justConcatenate = FALSE,
  trimOverhang = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{intermediate_path}{A path to the intermediate folder and directory}

\item{minOverlap}{(Optional). Default 12.
The minimum length of the overlap required for merging the forward and reverse reads.}

\item{maxMismatch}{(Optional). Default 0. 
The maximum mismatches allowed in the overlap region.}

\item{returnRejects}{(Optional). Default FALSE.
If TRUE, the pairs that that were rejected based on mismatches in the overlap
region are retained in the return \code{data.frame}.}

\item{justConcatenate}{(Optional). Default FALSE.
If TRUE, the forward and reverse-complemented reverse read are concatenated rather than merged,
  with a NNNNNNNNNN (10 Ns) spacer inserted between them.}

\item{trimOverhang}{(Optional). Default FALSE.
If TRUE, "overhangs" in the alignment between the forwards and reverse read are trimmed off.
"Overhangs" are when the reverse read extends past the start of the forward read, and vice-versa,
as can happen when reads are longer than the amplicon and read into the other-direction primer region.}

\item{verbose}{(Optional). Default FALSE. 
If TRUE, a summary of the function results are printed to standard output.}

\item{merged_read_data_path}{}
}
\description{
Merge forward and reverse reads ###CHECK
}
