% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.R
\name{minus}
\alias{minus}
\title{Minus one number from another}
\usage{
minus(x, y)
}
\arguments{
\item{x}{A numeric number}

\item{y}{Another numeric number}
}
\value{
x-y
}
\description{
Takes in two numbers. Minus one from another.
}
\examples{
minus(10,3)
7
minus(5,7)
-2
}
