% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Files_Primers.R
\name{orient_primers}
\alias{orient_primers}
\title{Take in user's primers and creates the complement, reverse, reverse complement of primers in one tibble}
\usage{
orient_primers(primer_path)
}
\arguments{
\item{primer_path}{a path to the csv file that holds the primer info for this project}
}
\value{
A tibble that contains the forward and reverse primers with all complements of primers
}
\description{
Take in user's primers and creates the complement, reverse, reverse complement of primers in one tibble
}
\examples{
primer_data <- orient_primers(primer_path). Part of the larger 'prepare_reads' function. 
}
