% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primer_hits.R
\name{post_primer_hit_data}
\alias{post_primer_hit_data}
\title{Get primer counts fo reach sample after primer removal and trimming steps}
\usage{
post_primer_hit_data(primer_data, cutadapt_data, intermediate_path)
}
\arguments{
\item{primer_data}{The primer data tibble created in prepare_primers function}

\item{cutadapt_data}{Intermediate_data folder with trimmed and filtered reads for each sample}

\item{intermediate_path}{A path to the intermediate folder and directory}
}
\description{
Get primer counts fo reach sample after primer removal and trimming steps
}
