% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trim_functions.R
\name{post_primer_hit_plot}
\alias{post_primer_hit_plot}
\title{Make a barplot of primers identified on reads after trim steps}
\usage{
post_primer_hit_plot(primer_hits, fastq_data, intermediate_path, plot_name)
}
\arguments{
\item{primer_hits}{A number of reads in which the primer is found}

\item{fastq_data}{A tibble with the fastq file paths, the direction of the sequences, and names of sequences}

\item{metadata}{A metadata containing the concatenated metadata and primer data}
}
\description{
Make a barplot of primers identified on reads after trim steps
}
