% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_prepfunctions.R
\name{post_trim_qc}
\alias{post_trim_qc}
\title{Wrapper script for plotQualityProfile after trim steps and primer removal.}
\usage{
post_trim_qc(cutadapt_data, intermediate_path, n = 5e+05)
}
\arguments{
\item{cutadapt_data}{Intermediate_data folder with trimmed and filtered reads for each sample}

\item{intermediate_path}{A path to the intermediate folder}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}
}
\description{
Wrapper script for plotQualityProfile after trim steps and primer removal.
}
