% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerpack.R
\docType{package}
\name{powerpack-package}
\alias{powerpack}
\alias{powerpack-package}
\title{powerpack: Adding squares abd cubes}
\description{
The powerpack package provides function that calculates
either squares or cubes of two numbers, then add
}
\section{Arbitary Sections}{

You can add arbitraty sections to your documentation,
}

\author{
\strong{Maintainer}: Concillia Mpofu \email{cm0789a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
