% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primer_hits.R
\name{pre_primer_hit_data}
\alias{pre_primer_hit_data}
\title{Get primer counts fo reach sample before primer removal and trimming steps
TODO: don't make if already}
\usage{
pre_primer_hit_data(primer_data, fastq_data, intermediate_path)
}
\arguments{
\item{primer_data}{The primer data tibble created in prepare_primers function}

\item{fastq_data}{A tibble with the fastq file paths, the direction of the sequences, and names of sequences}

\item{intermediate_path}{A path to the intermediate folder and directory}
}
\value{
A number of reads in which the primer is found
}
\description{
Get primer counts fo reach sample before primer removal and trimming steps
TODO: don't make if already
}
