% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{prepare}
\alias{prepare}
\title{Main command prepare reads for primer trimming}
\usage{
prepare(
  directory_path,
  primer_path,
  metadata_path,
  fastq_path,
  intermediate_path
)
}
\arguments{
\item{primer_path}{}

\item{fastq_path}{}

\item{intermediate_path}{}
}
\description{
Main command prepare reads for primer trimming
}
