% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareFiles.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Read in the metadata from user and combine it with the primer data}
\usage{
prepare_metadata(metadata_path, primer_data)
}
\arguments{
\item{metadata_path}{A path to the metadata .csv file}

\item{primer_data}{The primer data tibble created in prepare_primers function}
}
\value{
A metadata containing the concatenated metadata and primer data
}
\description{
Read in the metadata from user and combine it with the primer data
}
