% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Files_Primers.R
\name{primer_check}
\alias{primer_check}
\title{Matching Order Primer Check}
\usage{
primer_check(fastq_data)
}
\arguments{
\item{fastq_data}{A tibble with the fastq file paths, the direction of the sequences, and names of sequences}
}
\value{
None
}
\description{
Matching Order Primer Check
}
\examples{
Part of a larger function 'prepare_fastq' function. 
}
