% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cap.R
\name{print_cap}
\alias{print_cap}
\title{print_cap}
\usage{
print_cap(text)
}
\arguments{
\item{text}{A character string.}
}
\value{
A character string with capitalized letters.
}
\description{
This is a simple example function.
}
\examples{
print_cap("hello world")
}
