% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{process_rps10_barcode}
\alias{process_rps10_barcode}
\title{Process the information from an ASV abundance matrix to run DADA2 for rps10 barcode}
\usage{
process_rps10_barcode(returnList, asv_abund_matrix)
}
\arguments{
\item{returnList}{A list containing data modified by cutadapt, primer data, FASTQ data, and concatenated metadata and primer data}

\item{asv_abund_matrix}{An abundance matrix containing amplified sequence variants}
}
\description{
Process the information from an ASV abundance matrix to run DADA2 for rps10 barcode
}
