% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{process_rps10_ITS_barcode}
\alias{process_rps10_ITS_barcode}
\title{Main trim command to run core DADA2 functions for 2 barcodes}
\usage{
process_rps10_ITS_barcode(returnList, asv_abund_matrix)
}
\arguments{
\item{asv_abund_matrix}{An abundance matrix containing amplified sequence variants}

\item{returnListA}{list containing data modified by cutadapt, primer data, FASTQ data, and concatenated metadata and primer data}
}
\description{
Main trim command to run core DADA2 functions for 2 barcodes
}
