% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{remove_ns}
\alias{remove_ns}
\title{Wrapper function for core DADA2 filter and trim function for first filtering step}
\usage{
remove_ns(fastq_data, intermediate_path, maxN = 0, multithread = TRUE)
}
\arguments{
\item{fastq_data}{A tibble with the fastq file paths, the direction of the sequences, and names of sequences}

\item{intermediate_path}{A path to the intermediate folder}

\item{maxN}{(Optional). Default 0.
After truncation, sequences with more than \code{maxN} Ns will be discarded. 
Note that \code{\link[dada2]{dada}} does not allow Ns.}

\item{multithread}{(Optional). Default is FALSE.
 If TRUE, input files are filtered in parallel via \code{\link[parallel]{mclapply}}.
 If an integer is provided, it is passed to the \code{mc.cores} argument of \code{\link[parallel]{mclapply}}.
 Note that the parallelization here is by forking, and each process is loading another fastq file into
 memory. This option is ignored in Windows, as Windows does not support forking, with \code{mc.cores} set to 1.
If memory is an issue, execute in a clean environment and reduce the chunk size \code{n} and/or
 the number of threads.}

\item{metadata}{A metadata containing the concatenated metadata and primer data}
}
\description{
Wrapper function for core DADA2 filter and trim function for first filtering step
}
