% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{rps10_barcode_function}
\alias{rps10_barcode_function}
\title{Main trim command to run core DADA2 functions for rps10 barcode}
\usage{
rps10_barcode_function(intermediate_path, cutadapt_data)
}
\arguments{
\item{intermediate_path}{A path to the intermediate folder and directory}

\item{cutadapt_data}{Intermediate_data folder with trimmed and filtered reads for each sample}
}
\description{
Main trim command to run core DADA2 functions for rps10 barcode
}
