% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample}
\alias{sample}
\title{Wells in Nebraska with geographic coordinates}
\format{
sf
An sf/data.frame with 1,000 rows and 3 columns:
\describe{
  \item{wellid}{Well ID number}
  \item{gw_extracted}{Amount of groundwater extracted (made-up numbers)}
  \item{geometry}{Geometry of the wells}
}
}
\source{
<https://dnr.nebraska.gov/groundwater>
}
\usage{
sample
}
\description{
A subset of wells registered in Nebraska
}
\keyword{datasets}
