% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sum_plus_10}
\alias{sum_plus_10}
\title{Sum plus 10}
\usage{
sum_plus_10(num1, num2)
}
\arguments{
\item{num1}{Input integer number.}

\item{num2}{Input integer number.}
}
\value{
An integer number.
}
\description{
\code{sum_plus_10} sums 10 to a given number.
}
\examples{
sum_plus_10(3)
}
