% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_df.R
\name{unique_identifier}
\alias{unique_identifier}
\title{unique values in the data frame}
\usage{
unique_identifier(x)
}
\description{
This function identifies the the unique values and counts the length
}
\keyword{internal}
