% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_rows.R
\name{unique_rows}
\alias{unique_rows}
\title{unique_rows}
\usage{
unique_rows(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Numbers of unique values in each row, sorted from decreasing value
}
\description{
take a data frame and return how many unique values each row has
}
\examples{

data(mtcars)
unique_rows(mtcars)
}
