% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR_to_012.R
\name{vcfR_to_012}
\alias{vcfR_to_012}
\title{Convert vcfR to 012 format}
\usage{
vcfR_to_012(vcf)
}
\arguments{
\item{vcf}{vcf file}
}
\value{
vcf file with new GT format
}
\description{
Convert vcfR to 012 format
}
