% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareFiles.R
\name{create_intermediate}
\alias{create_intermediate}
\title{Create intermediate path and directory}
\usage{
create_intermediate(working_dir_path)
}
\arguments{
\item{working_dir_path}{A path to a directory containing reads and metadata/primer files}
}
\value{
A path to the intermediate folder and directory
}
\description{
Create intermediate path and directory
}
