% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareFiles.R
\name{create_intermediate}
\alias{create_intermediate}
\title{Create intermediate path and directory}
\usage{
create_intermediate(working_dir_path)
}
\arguments{
\item{working_dir_path}{}
}
\value{
intermediate_path
}
\description{
Create intermediate path and directory
}
