% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{format_abund_matrix}
\alias{format_abund_matrix}
\title{Format ASV abundance table}
\usage{
format_abund_matrix(asv_abund_table, seq_tax_asv)
}
\arguments{
\item{asv_abund_table}{An abundance matrix containing amplified sequence variants}

\item{seq_tax_asv}{An amplified sequence variants table with taxonomic information}
}
\description{
Format ASV abundance table
}
