% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(your_name)
}
\arguments{
\item{your_name}{Input string for a name of a person to say hello.}
}
\value{
A string saying hello.
}
\description{
\code{hello} says hello to a person.
}
\examples{
hello("Julia")
}
