% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{main_cutadapt_function}
\alias{main_cutadapt_function}
\title{Runs all of the functions associated with running cutadapt. Users should not have to touch these functions.}
\usage{
main_cutadapt_function(
  directory_path,
  primer_path,
  metadata_path,
  fastq_path,
  intermediate_path,
  cutadapt_path
)
}
\arguments{
\item{directory_path}{A path to a directory containing reads and metadata/primer files}

\item{primer_path}{The primer data tibble created in prepare_primers function}

\item{metadata_path}{A path to a metadata containing the concatenated metadata and primer data}

\item{fastq_path}{A path to a directory containing FASTQ reads}

\item{intermediate_path}{A path to the intermediate folder and directory}

\item{cutadapt_path}{A path to the cutadapt program}
}
\value{
Intermediate_data folder with trimmed and filtered reads for each sample.
}
\description{
Runs all of the functions associated with running cutadapt. Users should not have to touch these functions.
}
