% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareFiles.R
\name{matching_order_primer_check}
\alias{matching_order_primer_check}
\title{Matching Order Primer Check}
\usage{
matching_order_primer_check(fastq_data)
}
\arguments{
\item{fastq_data}{}
}
\value{
None
}
\description{
Matching Order Primer Check
}
