% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_prepfunctions.R
\name{plot_qc}
\alias{plot_qc}
\title{Wrapper function for plotQualityProfile function}
\usage{
plot_qc(cutadapt_data, intermediate_path, n = 5e+05)
}
\arguments{
\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}
}
\description{
Wrapper function for plotQualityProfile function
}
