% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primer_hits.R
\name{post_primer_hit_data}
\alias{post_primer_hit_data}
\title{Get primer counts fo reach sample after primer removal and trimming steps}
\usage{
post_primer_hit_data(primer_data, cutadapt_data, intermediate_path)
}
\arguments{
\item{intermediate_path}{}
}
\description{
Get primer counts fo reach sample after primer removal and trimming steps
}
