% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{prep_abund_table}
\alias{prep_abund_table}
\title{Prepare final ASV abundance matrix}
\usage{
prep_abund_table(cutadapt_data, asv_abund_table, locus)
}
\arguments{
\item{locus}{}
}
\description{
Prepare final ASV abundance matrix
}
