% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{prep_abund_table}
\alias{prep_abund_table}
\title{Prepare final ASV abundance matrix}
\usage{
prep_abund_table(cutadapt_data, asv_abund_table, locus)
}
\arguments{
\item{locus}{}

\item{Intermediate_data}{folder with trimmed and filtered reads for each sample}
}
\description{
Prepare final ASV abundance matrix
}
