% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primer_hits.R
\name{primer_hit_plot}
\alias{primer_hit_plot}
\title{Make a barplot of primers identified on reads}
\usage{
primer_hit_plot(primer_hits, fastq_data, intermediate_path, plot_name)
}
\arguments{
\item{plot_name}{}
}
\description{
Make a barplot of primers identified on reads
}
