% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sum_plus_10}
\alias{sum_plus_10}
\title{Sum plus 10}
\usage{
sum_plus_10(num)
}
\arguments{
\item{num}{Input number to be summed.}
}
\value{
An integer number.
}
\description{
\code{sum_plus_10} sums 10 to a given number.
}
\examples{
sum_plus_10(3)
}
